On the starting of services, reset the writers.
The writers seem to get themselves stuck and in a bad state when the nodes
shift statuses. Reset them as part of the node status shift.

Dont attempt to kill projections, but rather just clean them up in memory
such as removing them from the list.
TODO: Check for proper cleanup